backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Natural Language Processing
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
nltk==3.8.1
textdistance==4.6.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Validation & Serialization
pydantic==2.5.1
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# File Utilities
aiofiles==23.2.1
python-magic==0.4.27

# Logging & Monitoring
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1